refactor: separate KnowhowEditor into components and hooks
