+++ src/app/(main)/boards/knowhow/_components/KnowhowEditor.tsx
@@ -1,29 +1,23 @@
 "use client";
-
-import { useCallback, useRef, useState, useEffect, FormEventHandler } from "react";
-import "react-quill/dist/quill.snow.css";
-import dynamic from "next/dynamic";
+import { FormEventHandler, useEffect } from "react";
 import Button from "@/components/Button/Button";
-import ReactQuill, { ReactQuillProps } from "react-quill";
-import useKnowhowImageMutation from "@/stores/queries/knowhow/post/useKnowhowImageMutation";
-import useKnowhowMutation from "@/stores/queries/knowhow/post/useKnowhowMutation";
-import { TKnowhow } from "@/types/knowhow.type";
 import { useUserContext } from "@/provider/contexts/UserContext";
 import { useModal } from "@/provider/contexts/ModalContext";
 import { useRouter } from "next/navigation";
-import { Tables } from "@/types/supabase";
-import { MAX_CONTENT_LENGTH } from "@/app/(main)/boards/knowhow/_constants";
 import ErrorMessage from "@/app/(main)/boards/knowhow/_components/ErrorMessage";
 import LoadingSpinner from "@/components/Loading/LoadingSpinner";
+import ReactQuill, { ReactQuillProps } from "react-quill";
+import useFormValidation from "@/app/(main)/boards/knowhow/_hooks/useFormValidation";
+import useImageHandler from "@/app/(main)/boards/knowhow/_hooks/useImageHandler";
+import useEditorState from "@/app/(main)/boards/knowhow/_hooks/useEditorState";
+import { formats, getModules } from "@/app/(main)/boards/knowhow/_utils/quillUtils";
+import useFormHandlers from "@/app/(main)/boards/knowhow/_hooks/useFormHandlers";
+import { TKnowhow } from "@/types/knowhow.type";
+import dynamic from "next/dynamic";
 
 type ForwardedQuillComponent = ReactQuillProps & {
   forwardedRef: React.Ref<ReactQuill>;
 };
-
-type KnowhowEditorProps = {
-  previousContent?: TKnowhow;
-};
-
 const QuillNoSSRWrapper = dynamic<ForwardedQuillComponent>(
   async () => {
     const { default: QuillComponent } = await import("react-quill");
@@ -36,225 +30,23 @@ const QuillNoSSRWrapper = dynamic<ForwardedQuillComponent>(
   { ssr: false }
 );
 
-const getModules = (imageHandler: () => void) => ({
-  toolbar: {
-    container: [[{ header: [1, 2, 3, 4, 5, false] }], ["bold"], ["link", "image"]],
-    handlers: {
-      image: imageHandler
-    }
-  },
-  clipboard: {
-    matchVisual: false
-  },
-  history: {
-    delay: 2000,
-    maxStack: 500,
-    userOnly: true
-  }
-});
-
-const formats = [
-  "header",
-  "font",
-  "size",
-  "bold",
-  "italic",
-  "underline",
-  "strike",
-  "align",
-  "blockquote",
-  "list",
-  "bullet",
-  "script",
-  "indent",
-  "direction",
-  "background",
-  "color",
-  "link",
-  "image",
-  "video"
-];
-
+type KnowhowEditorProps = {
+  previousContent?: TKnowhow;
+};
 function KnowhowEditor({ previousContent }: KnowhowEditorProps) {
-  const router = useRouter();
-  const { user } = useUserContext();
-  const { open } = useModal();
-  const { addKnowhowImage } = useKnowhowImageMutation();
-  const { addKnowhow, isPostPending, updateKnowhow, isPatchPending } = useKnowhowMutation();
-  const quillRef = useRef<ReactQuill>(null);
-  const [editorTitle, setEditorTitle] = useState<string>(previousContent?.title || "");
-  const [editorContent, setEditorContent] = useState<string>(previousContent?.content || "");
-  const [errorMessage, setErrorMessage] = useState<{ title: string; content: string }>({ title: "", content: "" });
-  const [isUploadingImage, setIsUploadingImage] = useState<boolean>(false);
-  console.log(editorContent);
+  const { editorTitle, setEditorTitle, editorContent, setEditorContent, errorMessage, setErrorMessage, quillRef } =
+    useEditorState({ previousContent });
+  const { imageHandler, handlePaste } = useImageHandler(quillRef);
+  const { validateKnowhowForm } = useFormValidation(setErrorMessage);
+  const { handleOpenSubmitModal, handleCancel, isUploadingImage } = useFormHandlers(
+    quillRef,
+    editorTitle,
 
-  const handleImageUpload = async (file: File) => {
-    const reader = new FileReader();
-    reader.onload = async (e) => {
-      const base64Image = e.target?.result;
-      if (typeof base64Image === "string") {
-        const quillObj = quillRef.current?.getEditor();
-        const range = quillObj?.getSelection();
-        if (quillObj && range) {
-          const img = new Image();
-          img.src = base64Image;
-          img.onload = () => {
-            const { width, height } = img;
-            quillObj.insertEmbed(range.index, "image", base64Image);
-            quillObj.setSelection(range.index + 1, 0);
-
-            const imgElement = quillObj.root.querySelector(`img[src='${base64Image}']`);
-            if (imgElement) {
-              imgElement.setAttribute("width", `${width}`);
-              imgElement.setAttribute("height", `${height}`);
-            }
-          };
-        }
-      }
-      setIsUploadingImage(false);
-    };
-
-    reader.readAsDataURL(file);
-  };
-
-  const imageHandler = useCallback(() => {
-    const input = document.createElement("input");
-    input.setAttribute("type", "file");
-    input.setAttribute("accept", "image/*");
-    input.click();
-    input.onchange = () => {
-      const file = input.files ? input.files[0] : null;
-      if (file) {
-        handleImageUpload(file);
-      }
-    };
-  }, []);
-
-  const handlePaste = useCallback((e: ClipboardEvent) => {
-    const items = e.clipboardData?.items;
-    if (items) {
-      for (let i = 0; i < items.length; i++) {
-        if (items[i].type.startsWith("image/")) {
-          const file = items[i].getAsFile();
-          if (file) {
-            e.preventDefault();
-            handleImageUpload(file);
-          }
-        }
-      }
-    }
-  }, []);
-
-  const validateKnowhowForm = (newKnowhow: Partial<Tables<"knowhow_posts">>): boolean => {
-    const newErrorMessage = { title: "", content: "" };
-    const doc = new DOMParser().parseFromString(newKnowhow.content || "", "text/html");
-    const textContent = doc.body.textContent || "";
-
-    if (!newKnowhow.title?.trim()) {
-      newErrorMessage.title = "제목은 필수 입력 항목입니다.";
-    }
-    if (!textContent) {
-      newErrorMessage.content = "내용을 입력해주세요.";
-    }
-    if (textContent.length > MAX_CONTENT_LENGTH) {
-      newErrorMessage.content = `내용은 ${MAX_CONTENT_LENGTH}자 이하로 입력해주세요.`;
-    }
-    if (newErrorMessage.title || newErrorMessage.content) {
-      setErrorMessage(newErrorMessage);
-      return false;
-    }
-    return true;
-  };
-
-  const handleSubmit = async () => {
-    setIsUploadingImage(true);
-    let content = quillRef.current?.getEditor().root.innerHTML;
-
-    const base64Images = content?.match(/src="data:image\/[^"]+"/g) || [];
-    const urlImages = content?.match(/src="https:\/\/[^"]+"/g) || [];
-    const imageFiles: { file: File; base64: string; width: number; height: number }[] = [];
-    const imageUrls: string[] = [];
-
-    base64Images.forEach((imageString, index) => {
-      const base64Data = imageString.split('"')[1];
-      const file = dataURLtoFile(base64Data, `image-${Date.now()}-${index}.png`);
-      const imgElement = quillRef.current?.getEditor().root.querySelector(`img[src='${base64Data}']`);
-      const width = imgElement ? parseInt(imgElement.getAttribute("width") || "0") : 0;
-      const height = imgElement ? parseInt(imgElement.getAttribute("height") || "0") : 0;
-      imageFiles.push({ file, base64: base64Data, width, height });
-    });
-
-    urlImages.forEach((imageString) => {
-      const url = imageString.split('"')[1];
-      imageUrls.push(url);
-    });
-
-    if (imageFiles.length > 0) {
-      const formData = new FormData();
-      imageFiles.forEach(({ file }) => {
-        formData.append("img", file);
-      });
-
-      try {
-        const data = await addKnowhowImage(formData);
-
-        data.forEach((url: { publicUrl: string }, index: number) => {
-          const publicUrl = url.publicUrl;
-          const base64Data = imageFiles[index].base64;
-          const width = imageFiles[index].width;
-          const height = imageFiles[index].height;
-
-          content = content?.replace(base64Data, publicUrl);
-          content = content?.replace(
-            `<img src="${publicUrl}"`,
-            `<img src="${publicUrl}" width="${width}" height="${height}"`
-          );
-
-          imageUrls.push(publicUrl);
-        });
-      } catch (error) {
-        console.log(error);
-      }
-    }
-
-    if (user) {
-      const newKnowhow = {
-        title: editorTitle,
-        content,
-        image_urls: imageUrls,
-        user_id: user.userId
-      };
-      const isValid = validateKnowhowForm(newKnowhow);
-      if (isValid) {
-        if (previousContent) {
-          await updateKnowhow({ ...newKnowhow, knowhow_postId: previousContent.knowhow_postId });
-        } else {
-          await addKnowhow(newKnowhow);
-        }
-      }
-    }
-  };
-
-  const dataURLtoFile = (dataurl: string, filename: string) => {
-    const arr = dataurl.split(",");
-    const mime = arr[0].match(/:(.*?);/)![1];
-    const bstr = atob(arr[1]);
-    let n = bstr.length;
-    const u8arr = new Uint8Array(n);
-    while (n--) {
-      u8arr[n] = bstr.charCodeAt(n);
-    }
-    return new File([u8arr], filename, { type: mime });
-  };
+    previousContent,
+    validateKnowhowForm
+  );
 
   const modules = getModules(imageHandler);
-
-  useEffect(() => {
-    if (previousContent) {
-      setEditorTitle(previousContent.title);
-      setEditorContent(previousContent.content);
-    }
-  }, [previousContent]);
 
   useEffect(() => {
     const quillObj = quillRef.current?.getEditor();
@@ -265,23 +57,6 @@ function KnowhowEditor({ previousContent }: KnowhowEditorProps) {
       };
     }
   }, [handlePaste]);
-
-  const handleOpenSubmitModal: FormEventHandler<HTMLFormElement> = (e) => {
-    e.preventDefault();
-    open({
-      type: "confirm",
-      content: "글을 등록하시겠습니까?",
-      buttonContent: "등록",
-      onConfirm: handleSubmit
-    });
-  };
-
-  const handleCancel = () =>
-    open({
-      type: "confirm",
-      content: `정말 게시글 ${previousContent ? "수정" : "작성"}을 취소하시겠습니까?`,
-      onConfirm: () => router.back()
-    });
 
   return (
     <form onSubmit={handleOpenSubmitModal} className="h-full flex flex-col">
@@ -302,18 +77,16 @@ function KnowhowEditor({ previousContent }: KnowhowEditorProps) {
         value={editorContent}
         onChange={setEditorContent}
       />
-
       <ErrorMessage className="translate-y-11">{errorMessage?.content}</ErrorMessage>
-
       <div className="flex gap-[18px] self-end translate-y-16">
         <Button variant="white" size="medium" type="button" onClick={handleCancel}>
           취소하기
         </Button>
-        <Button variant="black" size="medium" type="submit" disabled={isPostPending || isUploadingImage}>
+        <Button variant="black" size="medium" type="submit" disabled={isUploadingImage}>
           등록하기
         </Button>
       </div>
-      {(isPostPending || isPatchPending || isUploadingImage) && <LoadingSpinner isSubmitting />}
+      {isUploadingImage && <LoadingSpinner isSubmitting />}
     </form>
   );
 }
